feat(Text): move to Content folder
