+++ packages/react-core/src/components/Content/examples/TextContentWrapped.tsx
