+++ packages/react-core/src/components/index.ts
@@ -17,6 +17,7 @@ export * from './Card';
 export * from './Checkbox';
 export * from './ClipboardCopy';
 export * from './CodeBlock';
+export * from './Content';
 export * from './DataList';
 export * from './DatePicker';
 export * from './DescriptionList';
@@ -62,7 +63,6 @@ export * from './Slider';
 export * from './Spinner';
 export * from './Switch';
 export * from './Tabs';
-export * from './Text';
 export * from './TextArea';
 export * from './TextInput';
 export * from './Tile';
